edit views and delete function in UserController

+++ app/Http/Controllers/UserController.php
@@ -714,13 +714,18 @@ class UserController extends Controller
      * @return view
      */
     public function deleteUser($user){
+        
         $user = User::find($user);
         $bank_card = BankCard::find($user->bank_card_id);
+
+        //dd($user->bank_card()->get());
+        if(!is_null($user)){
+            $user->delete();
+        }
         if(!is_null($bank_card)){
             $bank_card->delete();
-        }
-        $user->delete();  
-        return redirect()->route('users.list')->with('message',"L'utilisateur a bine été supprimé");
+        }  
+        return redirect()->route('users.list')->with('message',"L'utilisateur a bien été supprimé");
     }
